relocate import function
